create auth.test.js
