+++ src/__tests__/LoginButton.test.js
@@ -16,23 +16,21 @@ describe("LoginButton tests", () => {
   });
 
   test("login button returns correct action", () => {
-    useSession.mockReturnValueOnce([undefined, false]).mockReturnValueOnce([{user: {name: "username"}}, false]);
+    useSession.mockReturnValue([undefined, false]);
     render(<LoginButton/>);
 
     fireEvent.click(screen.getByRole("button", { name: "login" }));
 
     expect(signIn).toHaveBeenCalled();
-    expect(screen.getByRole("button", { name: "logout" })).toBeInTheDocument();
   });
 
   test("logout button returns correct action", () => {
-    useSession.mockReturnValueOnce([{user: {name: "username"}}, false]).mockReturnValueOnce([undefined, false]);
+    useSession.mockReturnValue([{user: {name: "username"}}, false]);
     render(<LoginButton/>);
 
-    fireEvent.click(screen.queryByRole("button", { name: "logout" }));
+    fireEvent.click(screen.getByRole("button", { name: "logout" }));
 
     expect(signOut).toHaveBeenCalled();
-    expect(screen.queryByRole("button", { name: "login" })).toBeInTheDocument();
   });
 
 });
