+++ src/__tests__/index.test.js
@@ -1,10 +1,14 @@
+/*
+
+Top-level integration Tests
+
+Any tests that require mocking the next-auth module should be put into auth.test.js
+
+*/
 import fetchMock from "fetch-mock-jest";
 import { render, screen, fireEvent } from "@testing-library/react";
 import { act } from "react-dom/test-utils";
 import WRMCWebsite from "../pages/index";
-import { useSession } from "next-auth/client";
-
-jest.mock("next-auth/client");
 
 describe("Top level integration tests", () => {
 
@@ -15,7 +19,7 @@ describe("Top level integration tests", () => {
   });
 
   test("Tests that the see-full-schedule button correctly displays the schedule", () => {
-    render(<WRMCWebsite />)
+    render(<WRMCWebsite />);
     fireEvent.click(screen.queryByRole("button", { name: "See Full Schedule" }));
     expect(screen.queryByTestId("schedule")).toBeInTheDocument();
   });
@@ -57,94 +61,6 @@ describe("Show details integration tests", () => {
     fireEvent.click(homePage);
     expect(screen.queryByTestId("show details page")).not.toBeInTheDocument();
     expect(screen.getByTestId("SOTW title")).toBeInTheDocument();
-  });
-});
-
-describe("Start show button integration tests", () => {
-  beforeEach(() => {
-    useSession.mockClear();
-  });
-
-  test("Start show button not visible when logged out", () => {
-    useSession.mockReturnValue([undefined, false]);
-    render(<WRMCWebsite />);
-    expect(screen.queryByRole("button", { name: "Start Show!" })).not.toBeInTheDocument();
-  });
-
-  test("Start show button visible when logged in", () => {
-    useSession.mockReturnValue([{user: {name: "username"}}, false]);
-    render(<WRMCWebsite />);
-    expect(screen.getByRole("button", { name: "Start Show!" })).toBeInTheDocument();
-  });
-
-  test("Start show button takes user to playlist logger", () => {
-    useSession.mockReturnValue([{user: {name: "username"}}, false]);
-    render(<WRMCWebsite />);
-    expect(screen.getByRole("button", { name: "Start Show!" })).toBeInTheDocument();
-    const options = screen.queryAllByTestId("show-option");
-    const selector = screen.getByRole("combobox");
-    fireEvent.change(selector, { target: { value: options[0].value }});
-    const startShowButton = screen.getByRole("button", { name: "Start Show!" });
-    expect(startShowButton).toBeEnabled();
-    fireEvent.click(startShowButton);
-    expect(screen.getByRole("button", { name: "Add Song" })).toBeInTheDocument();
-  });
-});
-
-describe.only("PlaylistLogger integration tests", () => {
-  const sampleTitle = "Sample Title";
-  const sampleArtist = "Sample Artist";
-  const sampleAlbum = "Sample Album";
-
-  const populateTextInputs = () => {
-    const titleInput = screen.getByRole("textbox", {name: "Title"});
-    const artistInput = screen.getByRole("textbox", {name: "Artist"});
-    const albumInput = screen.getByRole("textbox", {name: "Album"});
-
-    fireEvent.change(titleInput, { target: { value: sampleTitle } });
-    fireEvent.change(artistInput, { target: { value: sampleArtist } });
-    fireEvent.change(albumInput, { target: { value: sampleAlbum } });
-  }
-
-  beforeEach(() => {
-    useSession.mockClear();
-  });
-  
-  test("PlaylistLogger not visible when logged out", () => {
-    useSession.mockReturnValue([{user: {name: "username"}}, false]);
-    render(<WRMCWebsite />);
-    const options = screen.queryAllByTestId("show-option");
-    const selector = screen.getByRole("combobox");
-    fireEvent.change(selector, { target: { value: options[0].value }});
-    const startShowButton = screen.getByRole("button", { name: "Start Show!" })
-    fireEvent.click(startShowButton);
-    expect(screen.getByRole("button", { name: "Add Song" })).toBeInTheDocument();
-    fireEvent.click(screen.queryByRole("button", { name: "logout" }));
-    expect(screen.queryByRole("button", { name: "Add Song" })).not.toBeInTheDocument();
-    expect(screen.getByText("Show Of The Week")).toBeInTheDocument();
-  });
-
-  test("saved songs persist when navigating away from PlaylistLogger", () => {
-    useSession.mockReturnValue([{user: {name: "username"}}, false]);
-    render(<WRMCWebsite />);
-    const options = screen.queryAllByTestId("show-option");
-    const selector = screen.getByRole("combobox");
-    fireEvent.change(selector, { target: { value: options[0].value }});
-
-    fireEvent.click(screen.getByRole("button", { name: "Start Show!" }));
-    fireEvent.click(screen.getByRole("button", { name: "Add Song" }));
-
-    populateTextInputs()
-    fireEvent.click(screen.getByRole("button", { name: "Enter" }));
-    expect(screen.getByRole("button", { name: "Update" })).toBeInTheDocument();
-
-    fireEvent.click(screen.getByText("Home"));
-    fireEvent.click(screen.getByRole("button", { name: "Go to Current Playlist" }));
-
-    expect(screen.getByRole("textbox", {name: "Title"})).toHaveValue(sampleTitle);
-    expect(screen.getByRole("textbox", {name: "Artist"})).toHaveValue(sampleArtist);
-    expect(screen.getByRole("textbox", {name: "Album"})).toHaveValue(sampleAlbum);
-    expect(screen.getByRole("button", { name: "Update" })).toBeInTheDocument();
   });
 });
